# src/main/resources/META-INF/openapi.yaml
openapi: 3.0.3
info:
  description: Сервис предназначен для создания и обновления тарифов и пакетов услуг
  title: TariffConfigurator
  version: 1.0.0
  x-product-version: v1
  x-backend:
    name: TariffConfigurator
servers:
  - url: /tcr-web
paths:
  /tariff:
    get:
      operationId: tariffGet
      tags:
        - it-product-api
      summary: Метод получения тарифов по id
      parameters:
        - in: query
          name: id
          schema:
            type: string
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tariff'
    post:
      operationId: tariffAdd
      tags:
        - it-product-api
      summary: Метод добавления тарифов
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Tariff'
      responses:
        '200':
          description: OK
    put:
      operationId: tariffUpdate
      tags:
        - it-product-api
      summary: Метод обновления тарифов
      parameters:
        - in: query
          name: id
          schema:
            type: string
          required: true
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Tariff'
      responses:
        '200':
          description: OK
  /bundledProduct:
    get:
        operationId: BundleGet
        tags:
          - it-product-api
        summary: Метод получения пакетов по id
        parameters:
          - in: query
            name: id
            schema:
              type: string
            required: true
        responses:
          '200':
            description: OK
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/BundledProduct'
    post:
      operationId: BundleAdd
      tags:
        - it-product-api
      summary: Метод добавления пакетов
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BundledProduct'
      responses:
        '200':
          description: OK
    put:
      operationId: BundleUpdate
      tags:
        - it-product-api
      summary: Метод обновления пакетов
      parameters:
        - in: query
          name: id
          schema:
            type: string
          required: true
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BundledProduct'
      responses:
        '200':
          description: OK
  /search:
    get:
      operationId: search
      tags:
        - it-product-api
      summary: Поиск тарифов по критериям
      parameters:
        - in: query
          name: name
          schema:
            type: string
        - in: query
          name: UnlimInternet
          schema:
            type: boolean
        - in: query
          name: UnlimCalls
          schema:
            type: boolean
        - in: query
          name: archived
          schema:
            type: boolean
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tariff'

components:
  schemas:
    Tariff:
      example:
        archived: false
        deleted: false
        package:
          name: name
          description: description
        name: name
        id: id
        createDate: 2000-01-23T04:56:07.000+00:00
      properties:
        id:
          type: string
        name:
          maxLength: 128
          type: string
        createDate:
          format: date-time
          type: string
        archived:
          default: false
          type: boolean
        deleted:
          default: false
          type: boolean
        bundledProduct:
          $ref: '#/components/schemas/BundledProduct'
      required:
      - name
      title: Tariff
      type: object
    BundledProduct:
      example:
        name: name
        description: description
      properties:
        id:
          type: string
        name:
          maxLength: 128
          type: string
        createDate:
          format: date-time
          type: string
        products:
          $ref: '#/components/schemas/Products'
        deleted:
          default: false
          type: boolean
      required:
      - name
      - products
      title: BundledProduct
      type: object
    Products:
      properties:
        calls:
          type: integer
        sms:
          type: integer
        internet:
          type: integer
      title: Products
      type: object